Start work on List<BakedQuad>-based model impl
It's not possible to completely remove fabric-renderer-api usage
because it's not possible to make a BakedModel that's aware of the Level
it's in; that's a frapi extension. Forge provides similar utilities for
making level-aware models (see BakedModelWrapper) and it's
List<BakedQuad> based over there, so it will come in handy later.