Initial sketch of sticky packages (#13)
Putting a slimeblock or honey block next to a package makes it sticky.
Sticky packages remember items put into them and continue to filter
against then even when the contents drop below 0.

Removing the block will make it not sticky, the sticky stack will be
forgotten. This is not a property that is saved to the package item btw,
it's just on the block.

Manual insertion is currently buggy, gotta update the sticky stack. Also
want to add some more effects too